+++ pkg/nats/js_consumer_singleworker_push_group_handler.go
@@ -9,16 +9,15 @@ import (
 
 // jsPushTypeQueueGroupConsumer is a minimal Worker implementation that simply wraps a
 type jsConsumerPushQueueGroupSingeWorker struct {
-	subscriptionSrv subscriptionService
+	subscriptionSvc subscriptionService
 
-	handler func(msg *nats.Msg)
-	worker  *jsConsumerWorkerWrapper
+	worker *jsConsumerWorkerWrapper
 
 	logger *zap.Logger
 }
 
 func (wp *jsConsumerPushQueueGroupSingeWorker) OnReconnect(conn *nats.Conn) error {
-	err := wp.subscriptionSrv.OnReconnect(conn)
+	err := wp.subscriptionSvc.OnReconnect(conn)
 	if err != nil {
 		return err
 	}
@@ -27,7 +26,7 @@ func (wp *jsConsumerPushQueueGroupSingeWorker) OnReconnect(conn *nats.Conn) erro
 }
 
 func (wp *jsConsumerPushQueueGroupSingeWorker) OnDisconnect(conn *nats.Conn, err error) error {
-	retErr := wp.subscriptionSrv.OnDisconnect(conn, err)
+	retErr := wp.subscriptionSvc.OnDisconnect(conn, err)
 	if retErr != nil {
 		return retErr
 	}
@@ -35,8 +34,19 @@ func (wp *jsConsumerPushQueueGroupSingeWorker) OnDisconnect(conn *nats.Conn, err
 	return nil
 }
 
+func (wp *jsConsumerPushQueueGroupSingeWorker) OnClosed(conn *nats.Conn) error {
+	err := wp.subscriptionSvc.OnClosed(conn)
+	if err != nil {
+		wp.logger.Error("unable to call onClosed in consumer worker unit", zap.Error(err))
+	}
+
+	wp.subscriptionSvc = nil
+
+	return err
+}
+
 func (wp *jsConsumerPushQueueGroupSingeWorker) Init(ctx context.Context) error {
-	err := wp.subscriptionSrv.Init(ctx)
+	err := wp.subscriptionSvc.Init(ctx)
 	if err != nil {
 		return err
 	}
@@ -45,25 +55,23 @@ func (wp *jsConsumerPushQueueGroupSingeWorker) Init(ctx context.Context) error {
 }
 
 func (wp *jsConsumerPushQueueGroupSingeWorker) Healthcheck(ctx context.Context) bool {
-	return wp.subscriptionSrv.Healthcheck(ctx)
+	return wp.subscriptionSvc.Healthcheck(ctx)
 }
 
 func (wp *jsConsumerPushQueueGroupSingeWorker) Run(ctx context.Context) error {
-	err := wp.subscriptionSrv.Subscribe(ctx)
+	err := wp.subscriptionSvc.Subscribe(ctx)
 	if err != nil {
 		return err
 	}
 
-	return nil
-}
+	go func() {
+		<-ctx.Done()
 
-func (wp *jsConsumerPushQueueGroupSingeWorker) Shutdown(ctx context.Context) error {
-	err := wp.subscriptionSrv.Shutdown(ctx)
-	if err != nil {
-		return err
-	}
-
-	wp.handler = nil
+		err = wp.subscriptionSvc.UnSubscribe()
+		if err != nil {
+			wp.logger.Error("unable to unSubscribe", zap.Error(err))
+		}
+	}()
 
 	return nil
 }
@@ -79,7 +87,6 @@ func NewJsConsumerPushQueueGroupSingeWorker(logger *zap.Logger,
 
 	ww := &jsConsumerWorkerWrapper{
 		msgChannel:        nil, // cuz channel-less single-worker worker pool
-		stopWorkerChanel:  nil, // cuz channel-less single-worker worker pool
 		logger:            logger,
 		handler:           handler,
 		reQueueDelay:      requeueDelays,
@@ -90,7 +97,7 @@ func NewJsConsumerPushQueueGroupSingeWorker(logger *zap.Logger,
 
 	workersPool := &jsConsumerPushQueueGroupSingeWorker{
 		logger:          l,
-		subscriptionSrv: subscriptionSrv,
+		subscriptionSvc: subscriptionSrv,
 		worker:          ww,
 	}